Add A-Packages: duke package
