+++ src/main/java/duke/Storage.java
@@ -1,8 +1,8 @@
+package duke;
+
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Scanner;
 import java.nio.file.Files;
 import java.nio.file.Paths;
