+++ src/main/java/duke/ToDo.java
@@ -1,9 +1,10 @@
+package duke;
+
 /**
  * To do class IS-A Task.
  *
  * Task without any date/time attached to it
  * @author Timothy Wong Eu-Jin
- * @version A-MoreOOP
  */
 public class ToDo extends Task {
